// src/entities/Article/model/consts/articleConsts.ts
export enum ArticleSortField {
    TITLE = 'title',
    CREATED = 'createdAt',
    VIEW = 'views',
}

export enum ArticleBlockType {
    TEXT = 'TEXT',
    CODE = 'CODE',
    IMAGE = 'IMAGE'
}

export enum ArticleView {
    BIG = 'BIG',
    SMALL = 'SMALL',
}
export enum ArticleType {
    ALL = 'ALL',
    IT = 'IT',
    SCIENCE = 'SCIENCE',
    ECONOMICS = 'ECONOMICS'
}